pyproject.toml
[tool.poetry]
name = "pan-scm-cli"
version = "0.1.0"
description = "Network Engineer-friendly CLI for Palo Alto Networks Security Content Management"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "scm_cli", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
cmd2 = "^2.4.3"
rich = "^13.7.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
pan-scm-sdk = "^0.3.15"  # Actual SCM SDK package

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.6.1"
flake8 = "^6.1.0"

[tool.poetry.scripts]
scm-cli = "scm_cli.cli:main"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"